06-04-2022

+++ src/main/java/com/etiya/rentACar/business/requests/additionalServiceRequests/CreateAdditionalServiceRequest.java
@@ -14,5 +14,5 @@ public class CreateAdditionalServiceRequest {
 
     private String name;
 
-    private int dailyPrice;
+    private double dailyPrice;
 }
