+++ src/main/java/com/etiya/rentACar/business/requests/additionalServiceRequests/UpdateAdditionalServiceRequest.java
@@ -12,5 +12,5 @@ public class UpdateAdditionalServiceRequest {
 
     private String name;
 
-    private int dailyPrice;
+    private double dailyPrice;
 }
